+++ sdk/storage/azure-storage-file-datalake/azure/storage/filedatalake/__init__.py
@@ -27,6 +27,7 @@ from ._models import (
     UserDelegationKey,
     PublicAccess,
     AccessPolicy,
+    AccessControlChanges,
 )
 from ._shared_access_signature import generate_account_sas, generate_file_system_sas, generate_directory_sas, \
     generate_file_sas
@@ -58,6 +59,7 @@ __all__ = [
     'PathPropertiesPaged',
     'LeaseProperties',
     'ContentSettings',
+    'AccessControlChanges',
     'AccountSasPermissions',
     'FileSystemSasPermissions',
     'DirectorySasPermissions',