+++ sdk/storage/azure-storage-file-datalake/tests/test_directory.py
@@ -14,6 +14,7 @@ from azure.core.exceptions import HttpResponseError, ResourceExistsError, Resour
     ResourceModifiedError
 from azure.storage.filedatalake import ContentSettings, DirectorySasPermissions, DataLakeDirectoryClient
 from azure.storage.filedatalake import DataLakeServiceClient, generate_directory_sas
+from azure.storage.filedatalake._models import AccessControlChanges
 from testcase import (
     StorageTestCase,
     record,
@@ -23,6 +24,11 @@ from testcase import (
 
 # ------------------------------------------------------------------------------
 TEST_DIRECTORY_PREFIX = 'directory'
+REMOVE_ACL = "mask," + "default:user,default:group," + \
+             "user:ec3595d6-2c17-4696-8caa-7e139758d24a,group:ec3595d6-2c17-4696-8caa-7e139758d24a," + \
+             "default:user:ec3595d6-2c17-4696-8caa-7e139758d24a,default:group:ec3595d6-2c17-4696-8caa-7e139758d24a"
+
+
 # ------------------------------------------------------------------------------
 
 
@@ -64,9 +70,15 @@ class DirectoryTest(StorageTestCase):
         directory_client.create_directory()
         return directory_client
 
+    def _create_sub_directory_and_files(self, directory_client, num_of_dirs, num_of_files_per_dir):
+        # the name suffix matter since we need to avoid creating the same directories/files in record mode
+        for i in range(0, num_of_dirs):
+            sub_dir = directory_client.create_sub_directory(self.get_resource_name('subdir' + str(i)))
+            for j in range(0, num_of_files_per_dir):
+                sub_dir.create_file(self.get_resource_name('subfile' + str(j)))
+
     def _create_file_system(self):
         return self.dsc.create_file_system(self._get_file_system_reference())
-
 
     # --Helpers-----------------------------------------------------------------
 
@@ -189,7 +201,7 @@ class DirectoryTest(StorageTestCase):
 
         # to make sure the sub directory was indeed created by get sub_directory properties from sub directory client
         sub_directory_client = self.dsc.get_directory_client(self.file_system_name,
-                                                             directory_name+'/'+sub_directory_name)
+                                                             directory_name + '/' + sub_directory_name)
         sub_properties = sub_directory_client.get_directory_properties()
 
         # Assert
@@ -262,6 +274,323 @@ class DirectoryTest(StorageTestCase):
         # Assert
         self.assertIsNotNone(response)
         self.assertEquals(response['permissions'], 'rwxrwxrwx')
+
+    @record
+    def test_set_access_control_recursive(self):
+        directory_name = self._get_directory_reference()
+        directory_client = self.dsc.get_directory_client(self.file_system_name, directory_name)
+        directory_client.create_directory()
+        num_sub_dirs = 5
+        num_file_per_sub_dir = 5
+        self._create_sub_directory_and_files(directory_client, num_sub_dirs, num_file_per_sub_dir)
+
+        acl = 'user::rwx,group::r-x,other::rwx'
+        summary = directory_client.set_access_control_recursive(acl=acl)
+
+        # Assert
+        self.assertEqual(summary.directories_successful, num_sub_dirs + 1)  # +1 as the dir itself was also included
+        self.assertEqual(summary.files_successful, num_sub_dirs * num_file_per_sub_dir)
+        self.assertEqual(summary.failure_count, 0)
+        access_control = directory_client.get_access_control()
+        self.assertIsNotNone(access_control)
+        self.assertEqual(acl, access_control['acl'])
+
+    @record
+    def test_set_access_control_recursive_in_batches(self):
+        directory_name = self._get_directory_reference()
+        directory_client = self.dsc.get_directory_client(self.file_system_name, directory_name)
+        directory_client.create_directory()
+        num_sub_dirs = 5
+        num_file_per_sub_dir = 5
+        self._create_sub_directory_and_files(directory_client, num_sub_dirs, num_file_per_sub_dir)
+
+        acl = 'user::rwx,group::r-x,other::rwx'
+        summary = directory_client.set_access_control_recursive(acl=acl, batch_size=2)
+
+        # Assert
+        self.assertEqual(summary.directories_successful, num_sub_dirs + 1)  # +1 as the dir itself was also included
+        self.assertEqual(summary.files_successful, num_sub_dirs * num_file_per_sub_dir)
+        self.assertEqual(summary.failure_count, 0)
+        access_control = directory_client.get_access_control()
+        self.assertIsNotNone(access_control)
+        self.assertEqual(acl, access_control['acl'])
+
+    @record
+    def test_set_access_control_recursive_in_batches_with_progress_callback(self):
+        directory_name = self._get_directory_reference()
+        directory_client = self.dsc.get_directory_client(self.file_system_name, directory_name)
+        directory_client.create_directory()
+        num_sub_dirs = 5
+        num_file_per_sub_dir = 5
+        self._create_sub_directory_and_files(directory_client, num_sub_dirs, num_file_per_sub_dir)
+
+        acl = 'user::rwx,group::r-x,other::rwx'
+        running_tally = AccessControlChanges(0, 0, 0, "")
+
+        def progress_callback(resp):
+            running_tally.directories_successful += resp.directories_successful
+            running_tally.files_successful += resp.files_successful
+            running_tally.failure_count += resp.failure_count
+
+        summary = directory_client.set_access_control_recursive(acl=acl, progress_callback=progress_callback,
+                                                                batch_size=2)
+
+        # Assert
+        self.assertEqual(summary.directories_successful, num_sub_dirs + 1)  # +1 as the dir itself was also included
+        self.assertEqual(summary.files_successful, num_sub_dirs * num_file_per_sub_dir)
+        self.assertEqual(summary.failure_count, 0)
+        self.assertEqual(summary.directories_successful, running_tally.directories_successful)
+        self.assertEqual(summary.files_successful, running_tally.files_successful)
+        self.assertEqual(summary.failure_count, running_tally.failure_count)
+        access_control = directory_client.get_access_control()
+        self.assertIsNotNone(access_control)
+        self.assertEqual(acl, access_control['acl'])
+
+    @record
+    def test_set_access_control_recursive_with_failures(self):
+        root_directory_client = self.dsc.get_file_system_client(self.file_system_name)._get_root_directory_client()
+        root_directory_client.set_access_control(acl="user::--x,group::--x,other::--x")
+
+        # Using an AAD identity, create a directory to put files under that
+        directory_name = self._get_directory_reference()
+        token_credential = self.generate_oauth_token()
+        directory_client = DataLakeDirectoryClient(self.dsc.url, self.file_system_name, directory_name,
+                                                   credential=token_credential)
+        directory_client.create_directory()
+        num_sub_dirs = 5
+        num_file_per_sub_dir = 5
+        self._create_sub_directory_and_files(directory_client, num_sub_dirs, num_file_per_sub_dir)
+
+        # Create a file as super user
+        self.dsc.get_directory_client(self.file_system_name, directory_name).get_file_client("cannottouchthis") \
+            .create_file()
+
+        acl = 'user::rwx,group::r-x,other::rwx'
+        running_tally = AccessControlChanges(0, 0, 0, "")
+        failed_entries = []
+
+        def progress_callback(resp):
+            running_tally.directories_successful += resp.directories_successful
+            running_tally.files_successful += resp.files_successful
+            running_tally.failure_count += resp.failure_count
+            failed_entries.append(resp.failed_entries)
+
+        summary = directory_client.set_access_control_recursive(acl=acl, progress_callback=progress_callback,
+                                                                batch_size=2)
+
+        # Assert
+        self.assertEqual(summary.failure_count, 1)
+        self.assertEqual(summary.directories_successful, running_tally.directories_successful)
+        self.assertEqual(summary.files_successful, running_tally.files_successful)
+        self.assertEqual(summary.failure_count, running_tally.failure_count)
+        self.assertEqual(len(failed_entries), 1)
+
+    @record
+    def test_update_access_control_recursive(self):
+        directory_name = self._get_directory_reference()
+        directory_client = self.dsc.get_directory_client(self.file_system_name, directory_name)
+        directory_client.create_directory()
+        num_sub_dirs = 5
+        num_file_per_sub_dir = 5
+        self._create_sub_directory_and_files(directory_client, num_sub_dirs, num_file_per_sub_dir)
+
+        acl = 'user::rwx,group::r-x,other::rwx'
+        summary = directory_client.update_access_control_recursive(acl=acl)
+
+        # Assert
+        self.assertEqual(summary.directories_successful, num_sub_dirs + 1)  # +1 as the dir itself was also included
+        self.assertEqual(summary.files_successful, num_sub_dirs * num_file_per_sub_dir)
+        self.assertEqual(summary.failure_count, 0)
+        access_control = directory_client.get_access_control()
+        self.assertIsNotNone(access_control)
+        self.assertEqual(acl, access_control['acl'])
+
+    @record
+    def test_update_access_control_recursive_in_batches(self):
+        directory_name = self._get_directory_reference()
+        directory_client = self.dsc.get_directory_client(self.file_system_name, directory_name)
+        directory_client.create_directory()
+        num_sub_dirs = 5
+        num_file_per_sub_dir = 5
+        self._create_sub_directory_and_files(directory_client, num_sub_dirs, num_file_per_sub_dir)
+
+        acl = 'user::rwx,group::r-x,other::rwx'
+        summary = directory_client.update_access_control_recursive(acl=acl, batch_size=2)
+
+        # Assert
+        self.assertEqual(summary.directories_successful, num_sub_dirs + 1)  # +1 as the dir itself was also included
+        self.assertEqual(summary.files_successful, num_sub_dirs * num_file_per_sub_dir)
+        self.assertEqual(summary.failure_count, 0)
+        access_control = directory_client.get_access_control()
+        self.assertIsNotNone(access_control)
+        self.assertEqual(acl, access_control['acl'])
+
+    @record
+    def test_update_access_control_recursive_in_batches_with_progress_callback(self):
+        directory_name = self._get_directory_reference()
+        directory_client = self.dsc.get_directory_client(self.file_system_name, directory_name)
+        directory_client.create_directory()
+        num_sub_dirs = 5
+        num_file_per_sub_dir = 5
+        self._create_sub_directory_and_files(directory_client, num_sub_dirs, num_file_per_sub_dir)
+
+        acl = 'user::rwx,group::r-x,other::rwx'
+        running_tally = AccessControlChanges(0, 0, 0, "")
+
+        def progress_callback(resp):
+            running_tally.directories_successful += resp.directories_successful
+            running_tally.files_successful += resp.files_successful
+            running_tally.failure_count += resp.failure_count
+
+        summary = directory_client.update_access_control_recursive(acl=acl, progress_callback=progress_callback,
+                                                                   batch_size=2)
+
+        # Assert
+        self.assertEqual(summary.directories_successful, num_sub_dirs + 1)  # +1 as the dir itself was also included
+        self.assertEqual(summary.files_successful, num_sub_dirs * num_file_per_sub_dir)
+        self.assertEqual(summary.failure_count, 0)
+        self.assertEqual(summary.directories_successful, running_tally.directories_successful)
+        self.assertEqual(summary.files_successful, running_tally.files_successful)
+        self.assertEqual(summary.failure_count, running_tally.failure_count)
+        access_control = directory_client.get_access_control()
+        self.assertIsNotNone(access_control)
+        self.assertEqual(acl, access_control['acl'])
+
+    @record
+    def test_update_access_control_recursive_with_failures(self):
+        root_directory_client = self.dsc.get_file_system_client(self.file_system_name)._get_root_directory_client()
+        root_directory_client.set_access_control(acl="user::--x,group::--x,other::--x")
+
+        # Using an AAD identity, create a directory to put files under that
+        directory_name = self._get_directory_reference()
+        token_credential = self.generate_oauth_token()
+        directory_client = DataLakeDirectoryClient(self.dsc.url, self.file_system_name, directory_name,
+                                                   credential=token_credential)
+        directory_client.create_directory()
+        num_sub_dirs = 5
+        num_file_per_sub_dir = 5
+        self._create_sub_directory_and_files(directory_client, num_sub_dirs, num_file_per_sub_dir)
+
+        # Create a file as super user
+        self.dsc.get_directory_client(self.file_system_name, directory_name).get_file_client("cannottouchthis") \
+            .create_file()
+
+        acl = 'user::rwx,group::r-x,other::rwx'
+        running_tally = AccessControlChanges(0, 0, 0, "")
+        failed_entries = []
+
+        def progress_callback(resp):
+            running_tally.directories_successful += resp.directories_successful
+            running_tally.files_successful += resp.files_successful
+            running_tally.failure_count += resp.failure_count
+            failed_entries.append(resp.failed_entries)
+
+        summary = directory_client.update_access_control_recursive(acl=acl, progress_callback=progress_callback,
+                                                                   batch_size=2)
+
+        # Assert
+        self.assertEqual(summary.failure_count, 1)
+        self.assertEqual(summary.directories_successful, running_tally.directories_successful)
+        self.assertEqual(summary.files_successful, running_tally.files_successful)
+        self.assertEqual(summary.failure_count, running_tally.failure_count)
+        self.assertEqual(len(failed_entries), 1)
+
+    @record
+    def test_remove_access_control_recursive(self):
+        directory_name = self._get_directory_reference()
+        directory_client = self.dsc.get_directory_client(self.file_system_name, directory_name)
+        directory_client.create_directory()
+        num_sub_dirs = 5
+        num_file_per_sub_dir = 5
+        self._create_sub_directory_and_files(directory_client, num_sub_dirs, num_file_per_sub_dir)
+
+        summary = directory_client.remove_access_control_recursive(acl=REMOVE_ACL)
+
+        # Assert
+        self.assertEqual(summary.directories_successful, num_sub_dirs + 1)  # +1 as the dir itself was also included
+        self.assertEqual(summary.files_successful, num_sub_dirs * num_file_per_sub_dir)
+        self.assertEqual(summary.failure_count, 0)
+
+    @record
+    def test_remove_access_control_recursive_in_batches(self):
+        directory_name = self._get_directory_reference()
+        directory_client = self.dsc.get_directory_client(self.file_system_name, directory_name)
+        directory_client.create_directory()
+        num_sub_dirs = 5
+        num_file_per_sub_dir = 5
+        self._create_sub_directory_and_files(directory_client, num_sub_dirs, num_file_per_sub_dir)
+
+        summary = directory_client.remove_access_control_recursive(acl=REMOVE_ACL, batch_size=2)
+
+        # Assert
+        self.assertEqual(summary.directories_successful, num_sub_dirs + 1)  # +1 as the dir itself was also included
+        self.assertEqual(summary.files_successful, num_sub_dirs * num_file_per_sub_dir)
+        self.assertEqual(summary.failure_count, 0)
+
+    @record
+    def test_remove_access_control_recursive_in_batches_with_progress_callback(self):
+        directory_name = self._get_directory_reference()
+        directory_client = self.dsc.get_directory_client(self.file_system_name, directory_name)
+        directory_client.create_directory()
+        num_sub_dirs = 5
+        num_file_per_sub_dir = 5
+        self._create_sub_directory_and_files(directory_client, num_sub_dirs, num_file_per_sub_dir)
+
+        running_tally = AccessControlChanges(0, 0, 0, "")
+
+        def progress_callback(resp):
+            running_tally.directories_successful += resp.directories_successful
+            running_tally.files_successful += resp.files_successful
+            running_tally.failure_count += resp.failure_count
+
+        summary = directory_client.remove_access_control_recursive(acl=REMOVE_ACL, progress_callback=progress_callback,
+                                                                   batch_size=2)
+
+        # Assert
+        self.assertEqual(summary.directories_successful, num_sub_dirs + 1)  # +1 as the dir itself was also included
+        self.assertEqual(summary.files_successful, num_sub_dirs * num_file_per_sub_dir)
+        self.assertEqual(summary.failure_count, 0)
+        self.assertEqual(summary.directories_successful, running_tally.directories_successful)
+        self.assertEqual(summary.files_successful, running_tally.files_successful)
+        self.assertEqual(summary.failure_count, running_tally.failure_count)
+
+    @record
+    def test_remove_access_control_recursive_with_failures(self):
+        root_directory_client = self.dsc.get_file_system_client(self.file_system_name)._get_root_directory_client()
+        root_directory_client.set_access_control(acl="user::--x,group::--x,other::--x")
+
+        # Using an AAD identity, create a directory to put files under that
+        directory_name = self._get_directory_reference()
+        token_credential = self.generate_oauth_token()
+        directory_client = DataLakeDirectoryClient(self.dsc.url, self.file_system_name, directory_name,
+                                                   credential=token_credential)
+        directory_client.create_directory()
+        num_sub_dirs = 5
+        num_file_per_sub_dir = 5
+        self._create_sub_directory_and_files(directory_client, num_sub_dirs, num_file_per_sub_dir)
+
+        # Create a file as super user
+        self.dsc.get_directory_client(self.file_system_name, directory_name).get_file_client("cannottouchthis") \
+            .create_file()
+
+        running_tally = AccessControlChanges(0, 0, 0, "")
+        failed_entries = []
+
+        def progress_callback(resp):
+            running_tally.directories_successful += resp.directories_successful
+            running_tally.files_successful += resp.files_successful
+            running_tally.failure_count += resp.failure_count
+            failed_entries.append(resp.failed_entries)
+
+        summary = directory_client.remove_access_control_recursive(acl=REMOVE_ACL, progress_callback=progress_callback,
+                                                                   batch_size=2)
+
+        # Assert
+        self.assertEqual(summary.failure_count, 1)
+        self.assertEqual(summary.directories_successful, running_tally.directories_successful)
+        self.assertEqual(summary.files_successful, running_tally.files_successful)
+        self.assertEqual(summary.failure_count, running_tally.failure_count)
+        self.assertEqual(len(failed_entries), 1)
 
     @record
     def test_rename_from(self):
@@ -396,7 +725,7 @@ class DirectoryTest(StorageTestCase):
         dir1.create_sub_directory("subdir")
 
         dir2 = self._create_directory_and_get_directory_client("dir2")
-        dir2.rename_directory(dir1.file_system_name+'/'+dir1.path_name)
+        dir2.rename_directory(dir1.file_system_name + '/' + dir1.path_name)
 
         with self.assertRaises(HttpResponseError):
             dir2.get_directory_properties()